adjust `AggStateCache` module structure
