+++ src/stream/src/executor/aggregation/mod.rs
@@ -34,6 +34,7 @@ mod agg_call;
 mod agg_group;
 pub mod agg_impl;
 mod agg_state;
+mod agg_state_cache;
 mod distinct;
 mod minput;
 mod minput_agg_impl;
